Limpeza de warnings

+++ Controllers/MachinesController.cs
@@ -28,6 +28,10 @@ namespace NutriFitWeb.Controllers
         [Authorize(Roles = "client, trainer, nutritionist, gym")]
         public async Task<IActionResult> ShowMachines(string? searchString, string? currentFilter, int? pageNumber)
         {
+            if (User.Identity is null)
+            {
+                return BadRequest();
+            }
 
             if (searchString is not null)
             {
@@ -39,30 +43,31 @@ namespace NutriFitWeb.Controllers
             }
 
             HttpContext.Session.Clear();
+
             UserAccountModel user = await _userManager.FindByNameAsync(User.Identity.Name);
-            Gym gym = await _context.Gym.FirstOrDefaultAsync(a => a.UserAccountModel.Id == user.Id);
-            Trainer trainer = await _context.Trainer.Include(a => a.Gym).FirstOrDefaultAsync(a => a.UserAccountModel.Id == user.Id);
-            Nutritionist nutritionist = await _context.Nutritionist.Include(a => a.Gym).FirstOrDefaultAsync(a => a.UserAccountModel.Id == user.Id);
-            Client client = await _context.Client.Include(a => a.Gym).FirstOrDefaultAsync(a => a.UserAccountModel.Id == user.Id);
+            Gym? gym = await _context.Gym.FirstOrDefaultAsync(a => a.UserAccountModel.Id == user.Id);
+            Trainer? trainer = await _context.Trainer.Include(a => a.Gym).FirstOrDefaultAsync(a => a.UserAccountModel.Id == user.Id);
+            Nutritionist? nutritionist = await _context.Nutritionist.Include(a => a.Gym).FirstOrDefaultAsync(a => a.UserAccountModel.Id == user.Id);
+            Client? client = await _context.Client.Include(a => a.Gym).FirstOrDefaultAsync(a => a.UserAccountModel.Id == user.Id);
 
             ViewData["CurrentFilter"] = searchString;
             IQueryable<Machine>? machines = null;
 
             if (!string.IsNullOrEmpty(searchString) && gym is not null)
             {
-                machines = _context.Machines.Where(a => a.MachineGym == gym).Where(a => a.MachineName.Contains(searchString));
+                machines = _context.Machines.Where(a => a.MachineGym == gym).Where(a => a.MachineName != null && a.MachineName.Contains(searchString));
             }
             else if (!string.IsNullOrEmpty(searchString) && trainer is not null)
             {
-                machines = _context.Machines.Where(a => a.MachineGym == trainer.Gym).Where(a => a.MachineName.Contains(searchString));
+                machines = _context.Machines.Where(a => a.MachineGym == trainer.Gym).Where(a => a.MachineName != null && a.MachineName.Contains(searchString));
             }
             else if (!string.IsNullOrEmpty(searchString) && nutritionist is not null)
             {
-                machines = _context.Machines.Where(a => a.MachineGym == nutritionist.Gym).Where(a => a.MachineName.Contains(searchString));
+                machines = _context.Machines.Where(a => a.MachineGym == nutritionist.Gym).Where(a => a.MachineName != null && a.MachineName.Contains(searchString));
             }
             else if (!string.IsNullOrEmpty(searchString) && client is not null)
             {
-                machines = _context.Machines.Where(a => a.MachineGym == client.Gym).Where(a => a.MachineName.Contains(searchString));
+                machines = _context.Machines.Where(a => a.MachineGym == client.Gym).Where(a => a.MachineName != null && a.MachineName.Contains(searchString));
             }
             else if (gym is not null)
             {
@@ -81,8 +86,13 @@ namespace NutriFitWeb.Controllers
                 machines = _context.Machines.Where(a => a.MachineGym == client.Gym);
             }
 
-            int pageSize = 5;
-            return View(await PaginatedList<Machine>.CreateAsync(machines.AsNoTracking(), pageNumber ?? 1, pageSize));
+            if (machines is not null)
+            {
+                int pageSize = 5;
+                return View(await PaginatedList<Machine>.CreateAsync(machines.AsNoTracking(), pageNumber ?? 1, pageSize));
+            }
+
+            return NotFound();
         }
 
         [AllowAnonymous]
@@ -116,15 +126,15 @@ namespace NutriFitWeb.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> CreateMachinePost([Bind("MachineId, MachineProfilePhoto,MachineName,MachineDescription,MachineType,MachineQRCodeUri")] Machine machine, IFormFile? formFile)
         {
-            if (ModelState.IsValid)
+            if (ModelState.IsValid && User.Identity is not null)
             {
                 UserAccountModel user = await _userManager.FindByNameAsync(User.Identity.Name);
-                Gym gym = await _context.Gym.FirstOrDefaultAsync(a => a.UserAccountModel.Id == user.Id);
+                Gym? gym = await _context.Gym.FirstOrDefaultAsync(a => a.UserAccountModel.Id == user.Id);
 
 
                 if (gym is not null)
                 {
-                    List<Exercise> exercises = HttpContext.Session.Get<List<Exercise>>(SessionKeyExercises);
+                    List<Exercise>? exercises = HttpContext.Session.Get<List<Exercise>>(SessionKeyExercises);
                     machine.MachineExercises = exercises;
                     machine.MachineGym = gym;
                     machine.MachineProfilePhoto = _photoManagement.UploadProfilePhoto(formFile);
@@ -152,15 +162,15 @@ namespace NutriFitWeb.Controllers
                 .Include(a => a.MachineProfilePhoto)
                 .FirstOrDefaultAsync(a => a.MachineId == id);
 
-            if (machine is not null && machine.MachineProfilePhoto is not null)
+            if (machine is null)
+            {
+                return BadRequest();
+            }
+
+            if (machine.MachineProfilePhoto is not null)
             {
                 string? photoPath = _photoManagement.GetPhotoPath(machine.MachineProfilePhoto);
                 machine.MachineProfilePhoto.PhotoUrl = photoPath;
-            }
-
-            if (machine == null)
-            {
-                return NotFound();
             }
 
             HttpContext.Session.Set<List<Exercise>>(SessionKeyExercises, machine.MachineExercises);
@@ -172,37 +182,43 @@ namespace NutriFitWeb.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> EditMachinePost(int? id, IFormFile? formFile)
         {
-            if (id is null)
+            if (id is null || User.Identity is null)
+            {
+                return BadRequest();
+            }
+
+            Machine? machineToUpdate = await _context.Machines.Include(a => a.MachineExercises).FirstOrDefaultAsync(a => a.MachineId == id);
+
+            if (machineToUpdate is null)
             {
                 return NotFound();
             }
 
-            Machine? machineToUpdate = await _context.Machines.Include(a => a.MachineExercises).FirstOrDefaultAsync(a => a.MachineId == id);
             UserAccountModel user = await _userManager.FindByNameAsync(User.Identity.Name);
-            Gym gym = await _context.Gym.Include(a => a.Machines).FirstOrDefaultAsync(a => a.UserAccountModel.Id == user.Id);
+            Gym? gym = await _context.Gym.Include(a => a.Machines).FirstOrDefaultAsync(a => a.UserAccountModel.Id == user.Id);
 
-            if (gym is not null && machineToUpdate is not null && gym.Machines.Contains(machineToUpdate))
+            if (gym is not null && gym.Machines is not null && gym.Machines.Contains(machineToUpdate))
             {
                 Photo? oldPhoto = null;
-                if (machineToUpdate is not null && machineToUpdate.MachineProfilePhoto is not null)
+
+                if (machineToUpdate.MachineProfilePhoto is not null)
                 {
                     oldPhoto = machineToUpdate.MachineProfilePhoto;
                 }
-                if (machineToUpdate is not null)
-                {
-                    machineToUpdate.MachineProfilePhoto = _photoManagement.UploadProfilePhoto(formFile);
-                }
+                machineToUpdate.MachineProfilePhoto = _photoManagement.UploadProfilePhoto(formFile);
 
                 if (await TryUpdateModelAsync<Machine>(machineToUpdate, "",
-                    a => a.MachineName, a => a.MachineDescription, a => a.MachineType, a => a.MachineProfilePhoto))
+                    a => a.MachineName!, a => a.MachineDescription!, a => a.MachineType!, a => a.MachineProfilePhoto!))
                 {
-                    List<Exercise> exercises = HttpContext.Session.Get<List<Exercise>>(SessionKeyExercises);
+                    List<Exercise>? exercises = HttpContext.Session.Get<List<Exercise>>(SessionKeyExercises);
                     HttpContext.Session.Remove(SessionKeyExercises);
 
-                    HashSet<int>? excludedIDs = new(exercises.Select(a => a.ExerciseId));
-                    IEnumerable<Exercise>? missingRows = machineToUpdate.MachineExercises.Where(a => !excludedIDs.Contains(a.ExerciseId));
-
-                    _context.Exercise.RemoveRange(missingRows);
+                    if (exercises is not null && machineToUpdate.MachineExercises is not null)
+                    {
+                        HashSet<int>? excludedIDs = new(exercises.Select(a => a.ExerciseId));
+                        IEnumerable<Exercise>? missingRows = machineToUpdate.MachineExercises.Where(a => !excludedIDs.Contains(a.ExerciseId));
+                        _context.Exercise.RemoveRange(missingRows);
+                    }
 
                     machineToUpdate.MachineExercises = exercises;
 
@@ -239,13 +255,18 @@ namespace NutriFitWeb.Controllers
         [Authorize(Roles = "gym")]
         [HttpPost, ActionName("DeleteMachine")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteMachineConfirmed(int id)
+        public async Task<IActionResult> DeleteMachineConfirmed(int? id)
         {
+            if (id is null || User.Identity is null)
+            {
+                return BadRequest();
+            }
+
             Machine? machine = await _context.Machines.FindAsync(id);
             UserAccountModel user = await _userManager.FindByNameAsync(User.Identity.Name);
-            Gym gym = await _context.Gym.Include(a => a.Machines).FirstOrDefaultAsync(a => a.UserAccountModel.Id == user.Id);
+            Gym? gym = await _context.Gym.Include(a => a.Machines).FirstOrDefaultAsync(a => a.UserAccountModel.Id == user.Id);
 
-            if (gym is not null && machine is not null && gym.Machines.Contains(machine))
+            if (gym is not null && machine is not null && gym.Machines is not null && gym.Machines.Contains(machine))
             {
                 _context.Machines.Remove(machine);
                 await _context.SaveChangesAsync();
